+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [3.4.*]
+## [3.4.7] 2024-04-25
 - Added support for UBX-RXM-RTCM messages.
 - Added support for UBX-RXM-SPARTN messages.
 - Added new format option `lrf-uper` to output RTCM framed UPER encoded 3GPP LPP messages.